crucible: zero warnings on Clang 22 (-Weverything) and GCC 15 — 200+ hardening flags
Build system:
- Add -Weverything with strategic -Wno-* suppressions for Clang 22
- Add 170+ individual -W flags for GCC 15 (grouped by category)
- Add UBSan (undefined,integer) for Clang, (undefined,bounds-strict) for GCC
- Exclude unsigned-integer-overflow from sanitizers (hash/PRNG use intentional wrap)
- Remove GCC-incompatible -Wsfinae-incomplete=2 and -fsanitize=pointer-compare/subtract
- Remove noisy -Wsuggest-attribute=pure/const from GCC
- Raise -Wstack-usage to 512KB (background thread has 2MB stack)

Test fixes:
- Convert all raw uint64_t hash literals to SchemaHash/ShapeHash/ContentHash
- Replace old-style casts with static_cast throughout
- Fix double-promotion warnings in printf with static_cast<double>
- Fix float-equal in test_arena (range check instead of exact compare)
- Add [[maybe_unused]] for padding fields and intentionally unused constants
- Add (void) casts for intentionally discarded return values
- Add default: std::unreachable() to exhaustive switches

All 23 tests pass on both Clang 22 and GCC 15 with zero warnings.

+++ test/cpu_kernels.h
@@ -241,7 +241,8 @@ inline void sdpa(const float* Q, const float* K, const float* V,
 
     // Temp attention weights [S,S] — stack-allocate for small S.
     float stack_attn[128 * 128];
-    float* attn = (S <= 128) ? stack_attn : new float[S * S];
+    float* attn = (S <= 128) ? stack_attn
+        : new float[static_cast<size_t>(S) * static_cast<size_t>(S)];
 
     for (int b = 0; b < B; b++) {
         const float* q = Q + b * S * D;
